+++ DataAccessLayer/Repository/IPersonDataAccess.cs
@@ -2,10 +2,10 @@
 {
     public interface IPersonDataAccess
     {
-        long AddToken(long personID, string token);
-        long CheckToken(string token);
-        long GetPersonID(string username, string password);
-        long GetPersonIDForReadOnlyAccess(string username);
+        uint AddToken(long personID, string token);
+        uint CheckToken(string token);
+        uint GetPersonID(string username, string password);
+        uint GetPersonIDForReadOnlyAccess(string username);
         string GetPassword(string username, string password);
 
         string GetUserID(long personID);
