+++ DataAccessLayer/Repository/ITableTaskLinksDataAccess.cs
@@ -6,8 +6,8 @@ namespace DataAccessLayer
 {
     public interface ITableTasksLinksDataAccess
     {
-        long Insert(long personID, long[] taskIDs, string table_name, long entity_id);
-        IEnumerable<TableTaskLinks> Select(long personID, long[] taskIDs, string tableName, long entityID);
-        IEnumerable<TableTaskLinks> Select(long personID, long[] taskIDs, string tableName);
+        long Insert(uint personID, uint[] taskIDs, string table_name, uint entity_id);
+        IEnumerable<TableTaskLinks> Select(uint personID, uint[] taskIDs, string tableName, uint entityID);
+        IEnumerable<TableTaskLinks> Select(uint personID, uint[] taskIDs, string tableName);
     }
 }
