+++ WebAppApi48Core/Services/IAuthService.cs
@@ -6,9 +6,9 @@ namespace WebAppApi48Core.Services
 
     public interface IAuthService
     {
-        long VerifyCredentials(string userID, string password);
-        long CheckToken(string token);
+        uint VerifyCredentials(string userID, string password);
+        uint CheckToken(string token);
         string CreateToken(long personID, out string UserID, out string Token);
-        long GetPersonCode(HttpContext context);
+        uint GetPersonCode(HttpContext context);
     }
 }
